[m] Migrated most dependencies to conan packages
Only imgui and imgui-node-editor are left in the 'external' folder. All
the other dependencies are pulled and installed by cmake through conan
at configure time. Because the imgui node editor library and the
imgui docking branch are not available on the conan center they need
to be manually included inside the project. FetchContent, or
git submodules are'nt usable either, because neither project supports
cmake out of the box.

Added cmake-format config file, and added a clang-format file that
disables formatting for files located in 'external'.