[r] Reworked widget logic and graph editors
Added ability to register new widget implementations. Removed custom
panel classes for node editor and stylesheets. There is now only a
single Panel class, meant to act as a container for widgets, and be
dockable.

Removed the stylesheet mechanism entirely for now. Default style is fine
for the moment, and the old stylesheet implementation had a lot static
analysis warnings, polluting the build. Also removed old unused gui code
from the library, like the animation class and the imgui debug
utilities. The animation class might be reimplemented in the future,
I'm just trying to keep things slimmer and easier to understand for now.

Reimplemented the graph editor panel as widgets. There is now a graph
editor and a graph viewer. The graph viewer can make selections in the
graph, although they're not useful yet. Non-modifying operations on sets
of nodes will be added to a context menu as time goes on.

Reimplemented graph drawing code using Nelarius's imnodes library, which
seems like it has a simpler interface and is being actively maintained.

Made some quality of life tweaks to the ports. Inputports now have an
internal default output port which they connect to if not connected
to any external port. Connection queries still work as before, the
default port is treated as non existant from the outside.