[f] Add constant node types and refactor graph gui
Did a lot of changes for this feature, should have split it up into
multiple smaller merges.

The nodes are now conceptually separated from the algorithms
they run. A node works at graph level, and preoccupies itself
with graph topology, execution policies, maybe scheduling in
the future, and probably even more things. I have plans to add
nodes that enable templating mechanism, or that contain and
combine multiple sub-nodes in interesting ways. Separating
nodes from algorithms was essential open the project up
to this kind of pissibilities (haha I'm not correcting this).
Algorithms are simpler now, they really only provide a list of
inputs and a list of outputs, and act as a black box.

Already implemented a new kind of node, a constant node.
This node wraps no algorithm, but offers a dynamic number of
constant outputs. It can be created from the gui already, and you
can add 3 different types of constant ports to it. In the future I'll
delegate this to some factory function that knows all registered types.

In order to make the extension of the gui with customised node types
easier, I went ahead and split up the graph widgets into multiple
internal components. These are implementation details and shouldn't
affect client code at all, although I might expose it in the future if
I'm going to allow custom node types. There are still improvements
to be made, but it's easy enough to work with and extend for now.
I need to work on something other than gui code for a while.