[m] Rename the project to clayknot
Rendershop was already used, as I found out later. Also, as I time went
on I realised I don't want to limit the project to rendering. It can
do much more than that.

The new name sounds snappy, the namespace 'clk' is not worse than 'rsp',
and the name is vague enough so as to not grow outdated, but its very
search-able, and it's not in use as far as I could tell.