[r] Make node port getters return ranges
Although any_view incurrs a pretty big overhead on ranges that are
potentially going to be filtered and modified further, I think the
trade-off will be worth it for the flexibility and ease of use this will
offer. A lot of different node types are on the way, and they will
each store their ports in unpredictable ways. Also, having to collect
all ports in a container and return that is annoying.